requirements.txt
numpy>=1.20.0
pandas>=1.3.0
pillow>=8.0.0
opencv-python>=4.5.0
openslide-python>=1.1.2
tqdm>=4.60.0
